Fix a few compiler warnings and other minor issues (#535)
- sfio/sfdisc.c, vmalloc/malloc.c: Fix misleading indentation
  warnings and an unused variable warning.

- bltins/typeset.c: Avoid creating a new tree for tracked aliases
  if the only action required is to list the existing tracked
  aliases. This change improves the performance of 'alias -pt' when
  used in command substitutions by a slight amount.

- Ported minor fixes from graphviz (mostly for sign conversion):
  https://gitlab.com/graphviz/graphviz/-/commit/52558520
  https://gitlab.com/graphviz/graphviz/-/commit/b37ea3dc
  https://gitlab.com/graphviz/graphviz/-/commit/eebd1ab3
  https://gitlab.com/graphviz/graphviz/-/commit/f71883d4
  https://gitlab.com/graphviz/graphviz/-/commit/cf9c418e
  https://gitlab.com/graphviz/graphviz/-/commit/dcc5537c
  https://gitlab.com/graphviz/graphviz/-/commit/5a7f87c6
  https://gitlab.com/graphviz/graphviz/-/commit/d67cee19
  https://gitlab.com/graphviz/graphviz/-/commit/757513a0

- Remove the memory limits in sh_match.sh because those were
  causing the .sh.match tests to fail under ASan.
  Cf. https://github.com/att/ast/commit/1b1b53df

- GNU grep now issues a warning when run as egrep or fgrep:
  https://git.savannah.gnu.org/cgit/grep.git/commit/?id=a95156247
  To avoid these warnings, grep -E and grep -F are now used
  instead.

+++ src/cmd/ksh93/sh/io.c
@@ -2297,7 +2297,7 @@ struct eval
 {
 	Sfdisc_t	disc;
 	char		**argv;
-	short		slen;
+	int		slen;
 	char		addspace;
 };
 
@@ -2348,7 +2348,7 @@ static int eval_exceptf(Sfio_t *iop,int type, void *data, Sfdisc_t *handle)
 	if(!ep->addspace)
 	{
 		/* get the length of this string */
-		ep->slen = len = strlen(cp);
+		ep->slen = len = (int)strlen(cp);
 		/* move to next string */
 		ep->argv++;
 	}
